Added is_safe check and also known file extension check (in case a file is skipped) in the sort.py for Others folder, Removed used app_open variable and used app.state instead in back.py, Changed sorting trigger from size to modification time in back.py, Moved buttons to a menu in main.py, Removed add_config and remove_config and replaced with see_config function in settings.py and import in main.py, Made see_config to add and remove configs in settings.py, Fixed log file repitition, Made image paths relative to py or exe (pyinstaller), In is_safe function, Removed size check in sort.py, Updated modules used in README.md

+++ startup.py
@@ -1,32 +1,35 @@
 import os
 import sys
-import winreg
+import platform
+if platform.system() == "Windows":
+    import winreg
 
 def setup_startup(configs):
-    app_name = "File Cleaner"
-    script_path = os.path.join(os.path.dirname(os.path.abspath(__file__)), "main.py")
+    if platform.system() == "Windows":
+        app_name = "File Cleaner"
+        script_path = os.path.join(os.path.dirname(os.path.abspath(__file__)), "main.py")
 
-    python_exe = sys.executable.replace("python.exe", "pythonw.exe")
-    if not os.path.exists(python_exe):
-        python_exe = sys.executable
+        python_exe = sys.executable.replace("python.exe", "pythonw.exe")
+        if not os.path.exists(python_exe):
+            python_exe = sys.executable
 
-    key = winreg.OpenKey(
-        winreg.HKEY_CURRENT_USER,
-        r"Software\Microsoft\Windows\CurrentVersion\Run",
-        0,
-        winreg.KEY_SET_VALUE
-    )
+        key = winreg.OpenKey(
+            winreg.HKEY_CURRENT_USER,
+            r"Software\Microsoft\Windows\CurrentVersion\Run",
+            0,
+            winreg.KEY_SET_VALUE
+        )
 
-    try:
-        if configs.get("start"):
-            command = f'"{python_exe}" "{script_path}" -background'
-            winreg.SetValueEx(key, app_name, 0, winreg.REG_SZ, command)
-        else:
-            try:
-                winreg.DeleteValue(key, app_name)
-            except FileNotFoundError:
-                pass
-    except Exception as e:
-        print(e)
-    finally:
-        winreg.CloseKey(key)
+        try:
+            if configs.get("start"):
+                command = f'"{sys.executable}" -background' if hasattr(sys, "frozen") else f'"{python_exe}" "{script_path}" -background'
+                winreg.SetValueEx(key, app_name, 0, winreg.REG_SZ, command)
+            else:
+                try:
+                    winreg.DeleteValue(key, app_name)
+                except FileNotFoundError:
+                    pass
+        except Exception as e:
+            print(e)
+        finally:
+            winreg.CloseKey(key)
